[feature] Add routes for device pings

+++ server/models/mdb-models/phyllos-device-data.js
@@ -16,7 +16,9 @@ const deviceSchema = mongoose.Schema({
   noise       : []
 });
 
-deviceSchema.methods.record = (req, res) => { // req should be complete post body of device status
+deviceSchema.set('autoIndex', false);
+
+deviceSchema.statics.record = (req, res) => { // req should be {status: {DEVICE STATUS OBJECT}}
   Device.findOneAndUpdate({deviceId: req.deviceId}, {
     $set: {
       deviceId    : req.deviceId,
@@ -35,11 +37,25 @@ deviceSchema.methods.record = (req, res) => { // req should be complete post bod
     }
   }, {upsert: true}, (err) => {
     if (err) res.status(500).send(err);
-    res.status(200).send('Conditions for ', req.deviceId, ' recorded!');
+    res.status(200).send('Conditions for ' + req.deviceId + ' recorded!');
   });
 };
 
-// deviceSchema.methods.prune = (req, res) => { // req should be a device id string
+deviceSchema.statics.retrieve = (req, res) => {
+  Device.findOne({deviceId: req}).exec((err, device) => { // query with {deviceId: "id to be found"}
+    if (err) res.status(500).send(err);
+    res.status(200).send(device);
+  });
+};
+
+deviceSchema.statics.list = (req, res) => { // no arguments - returns all devices
+  Device.find((err, devices)=> {
+    if (err) res.status(500).send(err);
+    res.status(200).send(devices);
+  });
+};
+
+// deviceSchema.statics.prune = (req, res) => { // req should be a device id string
 //   Device.findOneAndUpdate({deviceId: req},
 //     {
 //       date        : {$slice:[]}, // Unsolved
@@ -57,43 +73,38 @@ deviceSchema.methods.record = (req, res) => { // req should be complete post bod
 //   });
 // };
 
-deviceSchema.methods.purge = (req, res) => { // req should be a device id string
-  Device.findOneAndUpdate({deviceId: req}, {
-    $set: {
-      date        : [],
-      moisture    : [],
-      ph          : [],
-      light       : [],
-      humidity    : [],
-      temperature : [],
-      pressure    : [],
-      noise       : []
-    }
-  }, {upsert: false}, (err) => {
-    if (err) res.status(500).send(err);
-    res.status(200).send('Conditions for ', req, ' recorded!');
-  });
+deviceSchema.statics.purge = (req, res) => { // query with {deviceId: "id to be found", confirm: "yes" or "no"}
+  if (req.confirm === "yes") {
+    Device.findOneAndUpdate({deviceId: req.deviceId}, {
+      $set: {
+        date        : [],
+        moisture    : [],
+        ph          : [],
+        light       : [],
+        humidity    : [],
+        temperature : [],
+        pressure    : [],
+        noise       : []
+      }
+    }, {upsert: false}, (err) => {
+      if (err) res.status(500).send(err);
+      res.status(200).send('Conditions for ' + req + ' recorded!');
+    });
+  } else {
+    res.status(300).send('You don\'t seem certain');
+  }
 };
 
-deviceSchema.methods.retrieve = (req, res) => {
-  Device.findOne({deviceId: req}).exec((err, device) => { // req should be a device id string
-    if (err) res.status(500).send(err);
-    res.status(200).send(device);
-  });
-};
-
-deviceSchema.methods.list = (req, res) => {
-  Device.find((err, devices)=> {
-    if (err) res.status(500).send(err);
-    res.status(200).send(devices);
-  });
-};
-
-deviceSchema.methods.discard = (req, res) => { // req should be a device id string
-  Device.findOneAndRemove({deviceId: req}, (err) => {
-    if (err) res.status(500).send(err);
-    res.status(200).send('Device ', req, ' has been removed from phyll.IO');
-  })
+deviceSchema.statics.discard = (req, res) => { // query with {deviceId: "id to be found", confirm: "yes" or "no"}
+  if (req.confirm === 'yes') {
+    Device.findOneAndRemove({deviceId: req}, (err) => {
+      if (err) res.status(500).send(err);
+      res.status(200).send('Device ' + req + ' has been removed from phyll.IO');
+    })
+  } else {
+    res.status(300).send('No take backs if you go through with this.');
+  }
 }
 
-module.exports = mongoose.model('Device', deviceSchema);
+const Device = mongoose.model('Device', deviceSchema);
+module.exports = Device;
